c2d::Audio: use float for fps

+++ include/cross2d/platforms/sdl2/sdl2_audio.h
@@ -14,7 +14,7 @@ namespace c2d {
 
     public:
 
-        SDL2Audio(int rate = 48000, int fps = 60, C2DAudioCallback cb = nullptr);
+        SDL2Audio(int rate = 48000, float fps = 60, C2DAudioCallback cb = nullptr);
 
         virtual ~SDL2Audio();
